Add workingDirectory support for launch profiles
`workingDirectory` was already modeled and resolved for
`commandName: "Executable"` profiles. I simply moved
`WorkingDirectory` and `TryParseWorkingDirectory` onto
the shared `LaunchProfile`/`LaunchProfileParser` base
types and parsed the value for Project profiles too.

The value is resolved relative to the directory
containing the launch settings file, matching the
existing Executable behavior.

An explicit `workingDirectory` takes precedence over
the `RunWorkingDirectory` MSBuild property, which
remains the fallback. Unlike `RunWorkingDirectory`,
which is per-project, this allows profiles of the
same project to launch in different directories.

Details can be found in #56128.

Fixes #56128
Fixes #20885
