+++ src/Cli/dotnet/Commands/Run/AotRunCommand.cs
@@ -141,7 +141,9 @@ internal static class AotRunCommand
                 applicationArguments,
                 profileResult.Profile?.CommandLineArgs,
                 appendApplicationArgumentsToBase: validatedRunProperties is not null);
-            workingDirectory = validatedRunProperties?.WorkingDirectory ?? currentDirectory;
+            workingDirectory = profileResult.Profile?.WorkingDirectory
+                ?? validatedRunProperties?.WorkingDirectory
+                ?? currentDirectory;
             runTier = plan.Tier;
             decisionReason = plan.Reason;
         }
