+++ src/Microsoft.DotNet.ProjectTools/LaunchSettings/ExecutableLaunchProfileParser.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Text.Json;
 
 namespace Microsoft.DotNet.ProjectTools;
@@ -38,30 +37,5 @@ internal sealed class ExecutableLaunchProfileParser : LaunchProfileParser
             DotNetRunMessages = profile.DotNetRunMessages,
             EnvironmentVariables = ParseEnvironmentVariables(profile.EnvironmentVariables),
         });
-    }
-
-    private static bool TryParseWorkingDirectory(string launchSettingsPath, string? value, out string? workingDirectory, [NotNullWhen(false)] out string? error)
-    {
-        if (value == null)
-        {
-            workingDirectory = null;
-            error = null;
-            return true;
-        }
-
-        var expandedValue = ExpandVariables(value);
-
-        try
-        {
-            workingDirectory = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(launchSettingsPath)!, expandedValue));
-            error = null;
-            return true;
-        }
-        catch
-        {
-            workingDirectory = null;
-            error = string.Format(Resources.Path0SpecifiedIn1IsInvalid, expandedValue, ExecutableLaunchProfile.WorkingDirectoryPropertyName);
-            return false;
-        }
     }
 }
