+++ src/Microsoft.DotNet.ProjectTools/LaunchSettings/LaunchProfileParser.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Immutable;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Microsoft.DotNet.ProjectTools;
 
@@ -11,6 +12,31 @@ internal abstract class LaunchProfileParser
 
     protected static string? ParseCommandLineArgs(string? value)
         => value != null ? ExpandVariables(value) : null;
+
+    protected static bool TryParseWorkingDirectory(string launchSettingsPath, string? value, out string? workingDirectory, [NotNullWhen(false)] out string? error)
+    {
+        if (value == null)
+        {
+            workingDirectory = null;
+            error = null;
+            return true;
+        }
+
+        var expandedValue = ExpandVariables(value);
+
+        try
+        {
+            workingDirectory = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(launchSettingsPath)!, expandedValue));
+            error = null;
+            return true;
+        }
+        catch
+        {
+            workingDirectory = null;
+            error = string.Format(Resources.Path0SpecifiedIn1IsInvalid, expandedValue, LaunchProfile.WorkingDirectoryPropertyName);
+            return false;
+        }
+    }
 
     public static string GetLaunchProfileDisplayName(string? launchProfile)
         => string.IsNullOrEmpty(launchProfile) ? Resources.DefaultLaunchProfileDisplayName : launchProfile;
